feat: پشتیبانی از مقدار اعشاری برای فیلد حق شیفت (PCT)
## مشکل
فیلد «حق شیفت» در حکم و قالب‌های حکم فقط مقادیر صحیح (مثلاً 7 یا 8 درصد)
را می‌پذیرفت و ورودی اعشاری مثل 7.5% امکان‌پذیر نبود. علاوه بر این، هنگام
ویرایش آیتم‌های غیرشیفت (مبالغ ریالی)، باگی وجود داشت که مبلغ را ضربدر 100
می‌کرد (مثلاً 5000000 به 500000000 تبدیل می‌شد).

## تغییرات

### پایگاه داده — نصب تازه (ScriptSqly.cs)
- PAY2_DECREE_LINE.AMOUNT: BIGINT → DECIMAL(18,2)
- PAY2_ITEM_TMPL_LINE.DEF_AMOUNT: BIGINT → DECIMAL(18,2)
- SP_PAY2_CALC_RUN: متغیر @ITEM_AMOUNT از BIGINT به DECIMAL(18,2)

### پایگاه داده — نصب‌های موجود (008_DecreeLineAmountDecimal.sql)
- Migration idempotent برای هر دو جدول با guard بررسی نوع bigint

### Program.cs
- اسکریپت 008 به لیست scriptsToRun اضافه شد

### DTO (Pay2EmployeeModels.cs)
- Pay2DecreeLineDto.AMOUNT: long → decimal
- Pay2ItemTmplLineDto.DEF_AMOUNT: long → decimal

### UI — DecreeLineModal.razor
- InputMode=Percentage و MaxDecimalPlaces=2 برای آیتم‌های شیفت PCT
- رفع باگ ×100 در EditLine: آیتم غیرشیفت با فرمت "0" (بدون اعشار) لود می‌شود
- UpdateShiftHint و SubmitSave: long.TryParse → decimal.TryParse با InvariantCulture
- GetMonthlyBaseSalary: نوع بازگشتی long? → decimal?

### UI — ItemTemplateLineModal.razor
- بارگذاری FullItemDefs و SHIFT_MODE (موازی با سایر درخواست‌ها)
- تشخیص آیتم SHIFT و سوئیچ خودکار InputMode به Percentage
- رفع باگ ×100 در EditLine: آیتم غیرشیفت با Truncate→long لود می‌شود
- OnItemChanged: ریست مبلغ هنگام تغییر آیتم انتخابی
- SubmitSave: long.TryParse → decimal.TryParse با InvariantCulture

### فایل‌های مرجع (به‌روزرسانی برای یکپارچگی)
- PAY2_DDL_v6.sql، PAY2_Procedures_v6.sql، 006_Pay2_HourlyCalcBasis.sql

+++ Server/Info/PAY2_DDL_v6.sql
@@ -522,7 +522,7 @@ CREATE TABLE [dbo].[PAY2_ITEM_TMPL_LINE]
 (
     [TMPL_ID]   INT      NOT NULL,
     [ITEM_ID]   INT      NOT NULL,
-    [DEF_AMOUNT] BIGINT  NOT NULL CONSTRAINT DF_TL_AMT DEFAULT(0),
+    [DEF_AMOUNT] DECIMAL(18,2) NOT NULL CONSTRAINT DF_TL_AMT DEFAULT(0),
     [INS_OV]    BIT      NULL,                                               -- NULL=از تعریف آیتم
     [TAX_OV]    BIT      NULL,
     [BASIS_OV]  TINYINT  NULL,
@@ -576,7 +576,7 @@ CREATE TABLE [dbo].[PAY2_DECREE_LINE]
 (
     [DEC_ID]   INT      NOT NULL,
     [ITEM_ID]  INT      NOT NULL,
-    [AMOUNT]   BIGINT   NOT NULL CONSTRAINT DF_DL_AMT DEFAULT(0),
+    [AMOUNT]   DECIMAL(18,2) NOT NULL CONSTRAINT DF_DL_AMT DEFAULT(0),
     [INS_OV]   BIT      NULL,                                                -- NULL=از PAY2_ITEM_DEF
     [TAX_OV]   BIT      NULL,
     [BASIS_OV] TINYINT  NULL,
